refactor: PointRedeemEvent 이름 변경, 필드 순서 변경 반영
